ISelectShiftsQuery - error fix

+++ Planday.Schedule/Queries/ISelectShiftsQuery.cs
@@ -6,7 +6,7 @@ namespace Planday.Schedule.Queries
     public interface ISelectShiftsQuery
     {
         Task<IReadOnlyCollection<Shift>> AllShifts();
-        Task<Shift> GetShiftById(int id);
+        Task<Shift> GetShiftById(long id);
     }    
 }
